LPS-112030 Move widgets functions to Utils
