+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/reducers/widgetsReducer.js
@@ -13,6 +13,8 @@
  */
 
 import {ADD_FRAGMENT_ENTRY_LINKS, DELETE_WIDGETS} from '../actions/types';
+import {getWidgetPath} from '../utils/getWidgetPath';
+import {setWidgetUsage} from '../utils/setWidgetUsage';
 
 export default function widgetsReducer(widgets, action) {
 	switch (action.type) {
@@ -66,80 +68,5 @@ export default function widgetsReducer(widgets, action) {
 
 		default:
 			return widgets;
-	}
-}
-
-/**
- * Get widget path from the widgets tree by portletId
- */
-function getWidgetPath(widgets, portletId, path = []) {
-	let widgetPath = null;
-
-	for (
-		let categoryIndex = 0;
-		categoryIndex < widgets.length;
-		categoryIndex += 1
-	) {
-		const {categories = [], portlets = []} = widgets[categoryIndex];
-
-		const categoryPortletIndex = portlets.findIndex(
-			(_portlet) => _portlet.portletId === portletId
-		);
-
-		const subCategoryPortletPath = getWidgetPath(categories, portletId, [
-			...path,
-			categoryIndex.toString(),
-			'categories',
-		]);
-
-		if (categoryPortletIndex !== -1) {
-			widgetPath = [
-				...path,
-				categoryIndex,
-				'portlets',
-				categoryPortletIndex,
-			];
-		}
-
-		if (subCategoryPortletPath) {
-			widgetPath = subCategoryPortletPath;
-		}
-	}
-
-	return widgetPath;
-}
-
-/**
- * Iterates the widgets array recursively to set a given widget 'used' property
- * and returns a new array with the given widget modified
- */
-function setWidgetUsage(widgets, path, usage) {
-	if (!path.length) {
-		return {
-			...widgets,
-			used: usage.used,
-		};
-	}
-
-	const [currentPath, ...restPath] = path;
-
-	if (Array.isArray(widgets)) {
-		return widgets.map((widget, index) => {
-			if (index === currentPath) {
-				return setWidgetUsage(widgets[currentPath], restPath, usage);
-			}
-
-			return widget;
-		});
-	}
-	else {
-		return {
-			...widgets,
-			[currentPath]: setWidgetUsage(
-				widgets[currentPath],
-				restPath,
-				usage
-			),
-		};
 	}
 }
